prereqs refactor
