--- lib/src/utils/prereqs/cabal/mod.rs
@@ -1,10 +1,0 @@
-pub mod check_installed_cabal;
-pub use check_installed_cabal::check_installed_cabal;
-pub mod check_cabal;
-pub use check_cabal::check_cabal;
-pub mod install_cabal;
-pub use install_cabal::install_cabal;
-pub mod get_cabal_version;
-pub use get_cabal_version::get_cabal_version;
-pub mod compare_cabal;
-pub use compare_cabal::compare_cabal;
