--- lib/src/utils/prereqs/ghc/mod.rs
@@ -1,12 +1,0 @@
-pub mod check_installed_ghc;
-pub use check_installed_ghc::check_installed_ghc;
-pub mod check_ghc;
-pub use check_ghc::check_ghc;
-pub mod install_ghc;
-pub use install_ghc::install_ghc;
-pub mod get_ghc_version;
-pub use get_ghc_version::get_ghc_version;
-pub mod compare_ghc;
-pub use compare_ghc::compare_ghc;
-pub mod get_ghc_version_url;
-pub use get_ghc_version_url::get_ghc_version_url;
