--- lib/src/utils/prereqs/ghcup/mod.rs
@@ -1,6 +1,0 @@
-pub mod check_ghcup;
-pub use check_ghcup::check_ghcup;
-pub mod install_ghcup;
-pub use install_ghcup::install_ghcup;
-pub mod get_ghcup_install_url;
-pub use get_ghcup_install_url::get_ghcup_install_url;
