--- lib/src/utils/prereqs/libsodium/mod.rs
@@ -1,6 +1,0 @@
-pub mod install_libsodium;
-pub use install_libsodium::install_libsodium;
-pub mod get_libsodium_url;
-pub use get_libsodium_url::get_libsodium_url;
-pub mod check_libsodium;
-pub use check_libsodium::check_libsodium;
